requirements.txt
# Core dependencies
python-dotenv==1.0.1
openai==1.59.4
pydantic==2.10.4
pydantic_core==2.27.2

# Vector store
upstash-redis==1.2.0
upstash-vector==0.1.1

# Data processing
pandas==2.1.0

# CLI and UI
halo==0.0.31
spinners==0.0.24
termcolor==2.5.0
colorama==0.4.6
tqdm==4.67.1

# HTTP and networking
aiohttp==3.11.11
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
httpcore==1.0.7
httpx==0.28.1
requests==2.32.3
urllib3==2.3.0

# Async support
anyio==4.8.0
sniffio==1.3.1

# Type checking and utilities
annotated-types==0.7.0
typing_extensions==4.12.2
PyYAML==6.0.2

# Other dependencies
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
jiter==0.8.2
log-symbols==0.0.14
multidict==6.1.0
propcache==0.2.1
six==1.17.0
yarl==1.18.3
